TG-78: Replace relative routes with absolute routes, created form folder and rename jsx files to index
